Move instance initializer cleanup to head-layout component
Previously, any eventual head data built by fastboot was cleaned up in an instance initializer. This was now moved into the head-layout component, where this happens on init.

This fixes an issue where this cleanup happens multiple times when using lazy loading engines.

Additionally, this moves the head-layout component into the addon namespace, which makes it possible to overwrite/extend from it.